Break dropdown menu into separate component
