+++ src/components/DropDown/DropDown.css
@@ -1,0 +1,13 @@
+.menu-wrapper {
+  z-index: 1000;
+  background-color: white;
+}
+
+.genre-button {
+  color: white;
+  z-index: 12;
+}
+
+.genre-menu {
+  color:white;
+}
